Fixed Rendered Conflict by Electron
The conflict caused by the frame when using rendered was caused as sidebar code was also using the same file name
